#100 - Undone routers names refactor and added modifications to the tests including the new :subject
