+++ gameservice/routers/RouterQuestionRetriever.test.js
@@ -2,7 +2,7 @@ const request = require('supertest');
 const express = require('express');
 const mongoose = require('mongoose');
 const Question = require('../question-model');
-const router = require('./QuestionProcessorRouter');
+const router = require('./RouterQuestionRetriever');
 
 const app = express();
 app.use(express.json());
@@ -23,14 +23,14 @@ describe('RouterQuestionRetriever', () => {
     });
     it('should return 3 random questions with 3 options each by default', async () => {
         await Question.insertMany([
-            { answer: 'Answer 1' },
-            { answer: 'Answer 2' },
-            { answer: 'Answer 3' },
-            { answer: 'Answer 4' },
-            { answer: 'Answer 5' }
+            { subject: 'Math', answer: 'Answer 1' },
+            { subject: 'Math', answer: 'Answer 2' },
+            { subject: 'Math', answer: 'Answer 3' },
+            { subject: 'Math', answer: 'Answer 4' },
+            { subject: 'Math', answer: 'Answer 5' }
         ]);
 
-        const res = await request(app).get('/api/game');
+        const res = await request(app).get('/api/game/Math');
 
         expect(res.status).toBe(200);
         expect(res.body.length).toBe(3);
@@ -42,15 +42,15 @@ describe('RouterQuestionRetriever', () => {
 
     it('should return the specified number of questions and options', async () => {
         await Question.insertMany([
-            { answer: 'Answer 1' },
-            { answer: 'Answer 2' },
-            { answer: 'Answer 3' },
-            { answer: 'Answer 4' },
-            { answer: 'Answer 5' },
-            { answer: 'Answer 6' }
+            { subject: 'Science', answer: 'Answer 1' },
+            { subject: 'Science', answer: 'Answer 2' },
+            { subject: 'Science', answer: 'Answer 3' },
+            { subject: 'Science', answer: 'Answer 4' },
+            { subject: 'Science', answer: 'Answer 5' },
+            { subject: 'Science', answer: 'Answer 6' }
         ]);
 
-        const res = await request(app).get('/api/game/2/4');
+        const res = await request(app).get('/api/game/Science/2/4');
 
         expect(res.status).toBe(200);
         expect(res.body.length).toBe(2);
@@ -62,10 +62,10 @@ describe('RouterQuestionRetriever', () => {
     
     it('should return 400 if not enough questions in DB', async () => {
         await Question.insertMany([
-            { answer: 'Answer 1' }
+            { subject: 'History', answer: 'Answer 1' }
         ]);
 
-        const res = await request(app).get('/api/game/3');
+        const res = await request(app).get('/api/game/History/3');
 
         expect(res.status).toBe(400);
         expect(res.body.error).toBe('Not enough questions in DB.');
@@ -76,7 +76,7 @@ describe('RouterQuestionRetriever', () => {
             throw new Error('Mocked error');
         });
 
-        const res = await request(app).get('/api/game');
+        const res = await request(app).get('/api/game/Math');
 
         expect(res.status).toBe(500);
         expect(res.body.error).toBe('Error retrieving questions');
